fakeintake: move PARTaskResult to api package
client/par.go was importing the server package solely to reference
PARTaskResult, which pulled prometheus (used by the server) as a
transitive dependency into test/e2e-framework and test/new-e2e modules.

Move PARTaskResult to the api package, following the existing pattern
for shared types (api.Payload, api.ParsedPayload). Both server and
client now import api instead, eliminating the client→server dependency
and the spurious prometheus entries from the e2e go.mod files.
